+++ Plugin_Shared/Assets/Plugins/SessionM/UserData.cs
@@ -9,7 +9,7 @@ public class UserData
 	private bool isOptedOut;
 	private bool isRegistered;
 	private bool isLoggedIn;
-	private long pointBalance;
+	private int pointBalance;
 	private long tierPointBalance;
 	private int unclaimedAchievementCount;
 	private int unclaimedAchievementValue;
@@ -22,7 +22,7 @@ public class UserData
 	private string startTier;
 
 	/*! Creates a new user with the specified parameters. */
-	public UserData(bool isOptedOut, bool isRegistered, bool isLoggedIn, long pointBalance, long tierPointBalance, int unclaimedAchievementCount, int unclaimedAcheivementValue, List<AchievementData> achievements, List<AchievementData> achievementsList, string tierIdentifier, string tierName, string tierPercentage, string tierAnniversaryDate, string startTier)
+	public UserData(bool isOptedOut, bool isRegistered, bool isLoggedIn, int pointBalance, long tierPointBalance, int unclaimedAchievementCount, int unclaimedAcheivementValue, List<AchievementData> achievements, List<AchievementData> achievementsList, string tierIdentifier, string tierName, string tierPercentage, string tierAnniversaryDate, string startTier)
 	{
 		this.isOptedOut = isOptedOut;
 		this.isRegistered = isRegistered;
@@ -47,7 +47,7 @@ public class UserData
 	/*! Returns whether user is currently logged in. */
 	public bool IsLoggedIn() { return this.isLoggedIn; }
 	/*! Returns user's current point balance. */
-	public long GetUserPointBalance() { return this.pointBalance; }
+	public int GetUserPointBalance() { return this.pointBalance; }
 	/*! Returns user's current tier point balance. */
 	public long GetUserTierPointBalance() { return this.tierPointBalance; }
 	/*! Returns how many achievements that the user has earned, but not claimed. */
